requirements.txt
matplotlib
numpy
pandas
scikit-learn
scipy
torch
torchvision
tensorboard
omegaconf
byol-pytorch
einops
torchxrayvision

cycler
decorator
imageio
kiwisolver
networkx
Pillow
pyparsing
python-dateutil
PyWavelets
scikit-image
six
opencv-python

torchviz
